+++ internal/proxy/freebind_other.go
@@ -1,0 +1,7 @@
+//go:build !linux
+
+package proxy
+
+func enableFreebind(fd uintptr) error {
+	return nil
+}